tabel and db chages

+++ src/app/routes/dashboard/dashboard.component.ts
@@ -6,18 +6,22 @@ import { Person } from '@config/model';
 import { peopleTableConfig } from '@config/table.config';
 import { generatedData } from '@db/db';
 import { NgxIndexedDBService } from 'ngx-indexed-db';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { ButtonModule } from 'primeng/button';
 import { Table, TableModule } from 'primeng/table';
 import { TooltipModule } from 'primeng/tooltip';
 import { ToastModule } from 'primeng/toast';
 import { AvatarModule } from 'primeng/avatar';
+import { TagModule } from 'primeng/tag';
+import { ConfirmDialogModule } from 'primeng/confirmdialog';
+import { format } from 'date-fns';
+import { CapitalizePipe } from '@pipes/captalize.pipe';
 
 @Component({
   selector: 'app-dashboard',
   standalone: true,
-  imports: [CommonModule, TableModule, ButtonModule, TooltipModule, FormsModule, ToastModule, AvatarModule],
-  providers: [MessageService],
+  imports: [CommonModule, TableModule, ButtonModule, TooltipModule, FormsModule, ToastModule, AvatarModule, TagModule, ConfirmDialogModule, CapitalizePipe],
+  providers: [MessageService, ConfirmationService],
   templateUrl: './dashboard.component.html',
   styleUrl: './dashboard.component.scss'
 })
@@ -25,6 +29,7 @@ export class DashboardComponent {
 
   private dbService = inject(NgxIndexedDBService);
   private messageService = inject(MessageService);
+  private confirmationService = inject(ConfirmationService);
   people: Person[] = [];
   tableConfig: any = peopleTableConfig;
   searchValue: string = '';
@@ -55,12 +60,12 @@ export class DashboardComponent {
     });
   }
 
-
   getUsers() {
     this.dbService.getAll(storeName).subscribe((result: any) => {
       this.people = result;
-      this.people.forEach((person: any) => {
-        person.fullName = `${person.firstName} ${person.middleName} ${person.lastName}`;
+      this.people.forEach((person: Person) => {
+        person.fullName = this.getFullName(person);
+        person.birthdate = person.birthdate ? format(person.birthdate, 'yyyy-MM-dd') : '';
         person.action = [
           { action: 'edit', iconClass: 'pi pi-pencil', tooltip: 'Edit', styleClass: 'text-blue-500' },
           { action: 'delete', iconClass: 'pi pi-trash', tooltip: 'Delete', styleClass: 'text-red-500' },
@@ -69,6 +74,10 @@ export class DashboardComponent {
       this.people = [...this.people.reverse()];
       console.log('result: ', this.people);
     });
+  }
+
+  getFullName(person: Person): string {
+    return `${person.firstName ? person.firstName : ''} ${person.middleName ? person.middleName : ''} ${person.lastName ? person.lastName : ''}`.trim();
   }
 
   addRecord() {
@@ -107,16 +116,26 @@ export class DashboardComponent {
   }
 
   onDelete(person: Person) {
-    if (!person.id) return;
-    this.dbService.delete(storeName, person.id).subscribe({
-      next: (allPeople) => {
-        console.log('Record deleted successfully.');
-        this.getUsers();
-        this.messageService.add({ severity: 'success', summary: 'Success', detail: 'Record deleted successfully.' });
-      },
-      error: (err: any) => {
-        console.log(err);
-        this.messageService.add({ severity: 'error', summary: 'Error', detail: err });
+    this.confirmationService.confirm({
+      message: `Are you sure that you want to delete user ${person.fullName}?`,
+      header: 'Confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      acceptIcon: "none",
+      rejectIcon: "none",
+      rejectButtonStyleClass: "p-button-text",
+      accept: () => {
+        if (!person.id) return;
+        this.dbService.delete(storeName, person.id).subscribe({
+          next: () => {
+            console.log('Record deleted successfully.');
+            this.getUsers();
+            this.messageService.add({ severity: 'success', summary: 'Success', detail: 'Record deleted successfully.' });
+          },
+          error: (err: any) => {
+            console.log(err);
+            this.messageService.add({ severity: 'error', summary: 'Error', detail: err });
+          }
+        });
       }
     });
   }
